Move to custom generated headers

+++ utilities/gfx-convert/intermediate/Tiles.hpp
@@ -19,7 +19,7 @@ class Tiles {
 public:
     Tiles(std::vector<uint8_t> image, uint16_t width, uint16_t height);
 
-    std::vector<uint8_t> ics_tiles();
+    std::vector<uint32_t> ics_tiles();
 
     std::vector<uint8_t> image;
     uint16_t width;
